Struct rusty_crc16::crc::Crc16

pub struct Crc16 {
    pub divisor: u16,
    pub initial_reminder: u16,
    pub table: Vec<u16>,
    pub reversed_input: bool,
    pub reverser_output: bool,
    pub xorout: u16,
}

Fields
divisor: u16 initial_reminder: u16 table: Vec<u16> reversed_input: bool reverser_output: bool xorout: u16
Methods

impl Crc16

fn create_table(&mut self)

Precomputes table for CRC calculation

Note: Calling calcuate CRC without having precomputed table will result in panic.

fn calculate(&self, bytes: &[u8]) -> u16

Calculates CRC16 value for given byte array

Arguments

bytes - A byte array whose value is being calculated.

fn calculate_rolling(&self, bytes: &[u8], current_value: u16) -> u16

Calculates CRC16 value for given byte array, but assumes calculation is resumed from previous state (multiple_parts)

Arguments

bytes - A byte array whose value is being calculated.

current_value - A CRC16 value obtained when calculating CRC16 of previous part.

fn check(&self, bytes: &[u8]) -> bool

Checks if byte array has valid CRC appended are continuing calculation from previous state (multiple_parts)

Arguments

bytes - A byte array whose value is being checked.

Note: Assuming that last two bytes in array are CRC16 value that was appended, this only works if whole data we are checking is in byte array. For larger files we can check validity by calculating CRC value again and checking if it is 0.